update https://github.com/d2-projects/d2-admin/pull/211
Former-commit-id: f246662bbf830bd098145c0a71d80f67259a9583
