+++ src/store/modules/d2admin/modules/gray.js
@@ -7,14 +7,14 @@ export default {
   mutations: {
     /**
      * @description 切换灰度状态
-     * @param {Object} state vuex state
+     * @param {Object} state state
      */
     toggle (state) {
       state.active = !state.active
     },
     /**
      * @description 设置灰度模式
-     * @param {Object} state vuex state
+     * @param {Object} state state
      * @param {Boolean} active active
      */
     set (state, active) {
