+++ src/store/modules/d2admin/modules/releases.js
@@ -5,7 +5,7 @@ export default {
   mutations: {
     /**
      * @description 显示版本信息
-     * @param {Object} state vuex state
+     * @param {Object} state state
      */
     versionShow () {
       util.log.capsule('D2Admin', `v${process.env.VUE_APP_VERSION}`)
